asim: move NewStateWith* helpers into separate file
Previously, NewStateWith* helpers were with simulator logic, even though
they were only used for unit tests (not datadriven tests). This commit
moves them into a separate file to make it clearer that they are purely testing
utilities and to clear confusion when changing code.

Epic: none
Release note: none
